[refactor] ChatMessageCreateRequest 패키지 이동 및 style, optimize imports
