+++ src/main/java/kr/co/pawong/pwbe/chat/application/port/in/CommandChatMessageDataUseCase.java
@@ -1,4 +1,5 @@
 package kr.co.pawong.pwbe.chat.application.port.in;
 
 public interface CommandChatMessageDataUseCase {
+
 }
