+++ src/main/java/kr/co/pawong/pwbe/chat/application/service/CommandChatMessageDataService.java
@@ -7,4 +7,5 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class CommandChatMessageDataService implements CommandChatMessageDataUseCase {
+
 }
